requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-telegram-bot==20.7
pymongo==4.6.0
python-dotenv==1.0.0
sympy==1.12
matplotlib==3.8.4
reportlab==4.0.9
Pillow==10.2.0
apscheduler==3.10.4
pytz==2023.4
numpy==1.26.4
requests==2.31.0
aiofiles==23.2.1
aiohttp==3.9.5
google-cloud-vision==3.4.5
google-generativeai==0.3.2
